devel-docs: fix automake when --disable-gtk-docs is used
A number of autmake files under devel-docs expand "CLEANFILES += ...". This
depends on gtkdocize setting up gtk-docs.make with (amongst other things)
"CLEANFILES = ...". However when running autogen.sh with --disable-gtk-docs,
gtk-docs.make is only populated with EXTRA_DIST - meaning CLEANFILES is not
set, and automake fails to parse the devel-docs automake files.

Therefore we need to explicitly set CLEANFILES when --disable-gtk-docs is
used to avoid seeing errors such as the following when running autogen:
  devel-docs/libgimp/Makefile.am:85: error: CLEANFILES must be set with '=' before using '+='

The CLEANFILES extension was first added in:
  0052803313 (devel-docs: REVERT THIS once module.actions get removed automatically, 2020-10-25)