desktop: fix duplication of the app icon in a dock like Plank
Without this entry, when starting GIMP as a flatpak application on
elementaryOS, the Dock icon would get duplicated, as if the application
was detached from its launcher. This entry fixes that by allowing the
dock to associate the app's window with the desktop entry.

More information on the type of problem this caused can be found here:
https://github.com/elementary/dock/issues/119