devel-docs: we must still restart a plug-in to debug with SIGCONT.
Commit 59f2ba44c7 changed the devel docs to tell that on Linux, just
running the gdb command `continue` is enough to restart the plug-in
process. Maybe it is true on some specific setups, so I leave this part
of the note, but for sure it never worked for me. I may run `continue`
as many times as I want, the process stays stopped until I explicitly
send a SIGCONT signal (our code raises itself a SIGSTOP which requires a
SIGCONT for continuation of the process).

Maybe in some configurations, gdb actually sends a SIGCONT when
`continue` is run but not in others? No idea.

Anyway I add back the part about sending a SIGCONT too, then people can
test and choose which procedure works for them.