plug-ins: fix static analysis warning.
I had completely forgotten about Coverity Scan but today received an
email with some new scan results. This was one of the warnings that we
had a possible resource leak.

In truth, this is mostly a false positive since gimp_export_image() will
flatten the image. Yet why not be thorough…